[web] Switch to doubles (flutter/engine#37336)

+++ engine/src/flutter/lib/web_ui/lib/src/engine/assets.dart
@@ -84,7 +84,7 @@ class AssetManager {
           printWarning('Asset manifest does not exist at `$url` – ignoring.');
           return Uint8List.fromList(utf8.encode('{}')).buffer.asByteData();
         }
-        throw AssetManagerException(url, request.status!);
+        throw AssetManagerException(url, request.status!.toInt());
       }
 
       final String? constructorName = target == null ? 'null' :
